re-org sharing utils.go
