--- api/common/sharing/validators.go
@@ -1,20 +1,0 @@
-package sharing
-
-import (
-	"github.com/Peripli/service-manager/pkg/types"
-	"github.com/Peripli/service-manager/pkg/util"
-	"github.com/Peripli/service-manager/pkg/web"
-	"github.com/tidwall/gjson"
-)
-
-func IsValidReferenceInstancePatchRequest(req *web.Request, instance *types.ServiceInstance, planIDProperty string) error {
-	newPlanID := gjson.GetBytes(req.Body, planIDProperty).String()
-	if instance.ServicePlanID != newPlanID {
-		return util.HandleInstanceSharingError(util.ErrChangingPlanOfReferenceInstance, instance.ID)
-	}
-	parametersRaw := gjson.GetBytes(req.Body, "parameters").Raw
-	if parametersRaw != "" {
-		return util.HandleInstanceSharingError(util.ErrChangingParametersOfReferenceInstance, instance.ID)
-	}
-	return nil
-}
